fix: missing legend
