+++ src/pages/config/3.tsx
@@ -6,7 +6,6 @@ import { RedirectButton } from '@/components/btns';
 import { TaoVotingBar } from '@/components/charts';
 import { useParams, useTaoVoting } from '@/hooks';
 import { TaoVotingTable } from '@/components/tables';
-import ChartLegend from '@/components/ChartLegend';
 
 function DisputableVoting() {
   const { barChart, table } = useTaoVoting();
@@ -21,44 +20,6 @@ function DisputableVoting() {
     submitProposal,
     handleChange,
   } = useParams();
-
-  const barChartLegend = [
-    {
-      name: 'non-quiet voting period',
-      bgColor: 'yellow',
-      tooltipText:
-        'The initial portion of the Vote Duration that will NOT trigger the Quiet Ending Extension',
-    },
-    {
-      name: 'delegated voting period',
-      bgColor: 'purple',
-      tooltipText:
-        'The amount of time delegates are permitted to vote on a proposal.',
-    },
-    {
-      name: 'normal vote duration',
-      bgColor: 'dark-blue',
-      tooltipText: 'The amount of time a proposal is eligible to be voted on.',
-    },
-    {
-      name: 'quiet ending period',
-      bgColor: 'orange',
-      tooltipText:
-        'If the voting outcome changes during this time the Quiet Ending Extension will trigger, extending the Vote Duration.',
-    },
-    {
-      name: 'quiet ending extension',
-      bgColor: 'blue',
-      tooltipText:
-        'The amount of time added to the Vote Duration resulting from the vote outcome changing during the Quiet Ending.',
-    },
-    {
-      name: 'execution delay',
-      bgColor: 'turquoise',
-      tooltipText:
-        'The amount of time after a vote passes before the proposed action is executed',
-    },
-  ];
 
   const inputs = [
     {
@@ -187,17 +148,6 @@ function DisputableVoting() {
                 barChart.proposalProcessWithExtension?.executionDelay
               }
             />
-            <div className="grid grid-rows-3 grid-flow-col text-gray pl-14 pb-6">
-              {barChartLegend.map((legend, index) => (
-                <ChartLegend
-                  key={index}
-                  name={legend.name}
-                  bgColor={legend.bgColor}
-                  colAlign
-                  tooltipText={legend.tooltipText}
-                />
-              ))}
-            </div>
             <TaoVotingTable table={table} />
           </ChartContainer>
         </div>
